style + db no longer required
